requirements.txt
cloudinary==1.2.0
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.9
django-cors-headers==1.1.0
django-debug-toolbar==1.4
django-extensions==1.5.9
django-filter==0.11.0
django-toolbelt==0.0.1
djangorestframework==3.3.2
geopy==1.11.0
gunicorn==19.4.1
httplib2==0.9.2
oauth2==1.9.0.post1
Pillow==3.0.0
psycopg2==2.6.1
PySocks==1.5.6
pytz==2015.7
requests==2.8.1
sendgrid==1.5.19
six==1.10.0
smtpapi==0.2.0
sqlparse==0.1.18
static3==0.6.1
twilio==5.1.0
wheel==0.24.0
